refactor add_unregister_command out
