+++ tests/cli/test_parsers.py
@@ -2,8 +2,8 @@ from unittest import TestCase
 from mock import Mock, call
 
 from tests import conditionally_return
-from lintreview.cli.parsers import add_register_command
-from lintreview.cli.handlers import register_hook
+from lintreview.cli.parsers import add_register_command, add_unregister_command
+from lintreview.cli.handlers import register_hook, remove_hook
 
 class TestCliParsers(TestCase):
 
@@ -47,3 +47,39 @@ class TestCliParsers(TestCase):
 
         register_command_parser.add_argument.assert_has_calls(expected_add_argument_calls)
         register_command_parser.set_defaults.assert_has_calls(expected_set_defaults_calls)
+
+    def test_add_unregister_command_adds_remove_hook_invocation(self):
+        commands = Mock()
+        remove_command_parser = Mock()
+
+        desc = "Unregister webhooks for a given user & repo."
+
+        commands.add_parser = conditionally_return(remove_command_parser, 'unregister', help=desc)
+
+        expected_add_argument_calls = [
+            call(
+                '-u', '--user',
+                dest='login_user',
+                help="The OAuth token of the user that has admin rights to the repo "
+                    "you are removing hooks from. Useful when the "
+                    "user in settings is not the administrator of "
+                    "your repositories."
+            ),
+            call(
+                'user',
+                help="The user or organization the repo is under."
+            ),
+            call(
+                'repo',
+                help="The repository to remove a hook from."
+            )
+        ]
+
+        expected_set_defaults_calls = [
+            call(func=remove_hook)
+        ]
+
+        add_unregister_command(commands)
+
+        remove_command_parser.add_argument.assert_has_calls(expected_add_argument_calls)
+        remove_command_parser.set_defaults.assert_has_calls(expected_set_defaults_calls)
